processor/requirements.txt
accelerate==1.11.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
colorlog==6.10.1
dill==0.4.0
docling==2.61.2
docling-core==2.51.0
docling-ibm-models==3.10.2
docling-parse==4.7.1
et_xmlfile==2.0.0
Faker==38.0.0
filelock==3.20.0
filetype==1.2.0
fsspec==2025.10.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
instructor==1.3.3
Jinja2==3.1.6
json-repair==0.30.2
jsonlines==4.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
latex2mathml==3.78.1
lxml==6.0.2
markdown-it-py==4.0.0
marko==2.2.1
MarkupSafe==3.0.3
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.18
networkx==3.5
numpy==2.2.6
ocrmac==1.0.0
ollama==0.6.1
omegaconf==2.3.0
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pluggy==1.6.0
polyfactory==2.22.4
psutil==7.1.3
pyclipper==1.3.0.post6
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pylatexenc==2.10
pyobjc-core==12.0
pyobjc-framework-Cocoa==12.0
pyobjc-framework-CoreML==12.0
pyobjc-framework-Quartz==12.0
pyobjc-framework-Vision==12.0
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
rapidocr==3.4.2
referencing==0.37.0
regex==2025.11.3
reportlab==4.4.4
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
rtree==1.4.1
safetensors==0.6.2
scipy==1.16.3
semchunk==2.2.2
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
soupsieve==2.8
sympy==1.14.0
tabulate==0.9.0
tokenizers==0.22.1
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.1
tree-sitter==0.23.2
tree-sitter-c==0.23.6
tree-sitter-java==0.23.5
tree-sitter-javascript==0.23.1
tree-sitter-python==0.23.6
tree-sitter-typescript==0.23.2
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xlsxwriter==3.2.9
